requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-chess>=1.10.0
aiohttp>=3.9.0
pydantic>=2.7.0
python-dotenv>=1.0.0
requests>=2.31.0
celery>=5.3.6
redis>=5.0.1
fastapi-socketio>=0.0.10
cachetools>=5.3.2

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.1
httpx>=0.27.0  # For FastAPI TestClient
pytest-asyncio==0.21.1
# httpx==0.24.1  # For FastAPI TestClient